requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
auto-py-to-exe==2.7.8
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.3
django-allauth==0.43.0
django-bootstrap3==14.2.0
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
Eel==0.12.4
filelock==3.0.12
future==0.18.2
gevent==20.9.0
gevent-websocket==0.10.1
greenlet==1.0a1
gunicorn==20.0.4
idna==2.10
isodate==0.6.0
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
macholib==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.22
mysqlclient==2.0.1
oauthlib==3.1.0
openapi-codec==1.3.2
pbr==5.5.1
pip-autoremove==0.9.1
protobuf==3.14.0
psycopg2-binary==2.8.6
pycparser==2.20
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.10
PyJWT==1.7.1
pylint==2.6.0
PyMySQL==0.10.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
static3==0.7.0
stevedore==3.2.2
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
uWSGI==2.0.19.1
virtualenv==20.1.0
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
whichcraft==0.6.1
whitenoise==5.2.0
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.2.0
